requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
cairosvg>=2.7.0
numpy>=1.24.0
elasticsearch>=8.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0

